mrc-remote-instance-weight-transport(rwt-migrate-register-bootstrap-move,mechanical_provable): Move _register_to_engine_info_bootstrap onto RemoteInstanceWeightTransport (cut+paste)
