+++ python/sglang/srt/model_executor/model_runner.py
@@ -673,9 +673,7 @@ class ModelRunner(ModelRunnerKVCacheMixin):
             self.remote_instance_weight_transport.weight_info = register_memory_region(
                 self.model, self.remote_instance_weight_transport.engine
             )
-            ModelRunner._register_to_engine_info_bootstrap(
-                self.remote_instance_weight_transport
-            )
+            self.remote_instance_weight_transport._register_to_engine_info_bootstrap()
 
         # For MTP models like DeepSeek-V3 or GLM-4.5, the MTP layer(s) are used separately as draft
         # models for speculative decoding. In those cases, `num_nextn_predict_layers` is used to
@@ -1003,53 +1001,6 @@ class ModelRunner(ModelRunnerKVCacheMixin):
                     "set_dspark_layers_to_capture nor set_dflash_layers_to_capture, "
                     "one of which is required for DFLASH/DSPARK."
                 )
-
-    @staticmethod
-    def _register_to_engine_info_bootstrap(self: RemoteInstanceWeightTransport):
-        """Register transfer engine info with the EngineInfoBootstrapServer via HTTP PUT.
-
-        The bootstrap server runs on node_rank==0. For multi-node setups, the
-        host is derived from dist_init_addr. For single-node, use 127.0.0.1.
-        """
-        import requests as http_requests
-
-        if self.server_args.dist_init_addr:
-            # Multi-node: bootstrap server is on the head node (node_rank==0).
-            # Derive host from dist_init_addr (shared across all nodes).
-            bootstrap_host = (
-                NetworkAddress.parse(self.server_args.dist_init_addr).resolved().host
-            )
-        else:
-            bootstrap_host = "127.0.0.1"
-
-        bootstrap_port = self.server_args.engine_info_bootstrap_port
-        bootstrap_na = NetworkAddress(bootstrap_host, bootstrap_port)
-        url = f"{bootstrap_na.to_url()}/register_transfer_engine_info"
-
-        payload = {
-            "tp_rank": self.tp_rank,
-            "transfer_engine_info": {
-                "session_id": self.session_id,
-                "weights_info_dict": self.weight_info,
-            },
-        }
-
-        try:
-            resp = http_requests.put(url, json=payload, timeout=5)
-            if resp.status_code == 200:
-                logger.info(
-                    f"Registered transfer engine info for tp_rank={self.tp_rank} "
-                    f"with bootstrap server at {bootstrap_na}"
-                )
-            else:
-                logger.error(
-                    f"Failed to register transfer engine info for tp_rank={self.tp_rank}: "
-                    f"{resp.status_code}, {resp.text}"
-                )
-        except Exception as e:
-            logger.error(
-                f"Failed to register transfer engine info for tp_rank={self.tp_rank}: {e}"
-            )
 
     def check_quantized_moe_compatibility(self):
         if (
